Linux build fix #3 (RPM)
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@43719 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Installer/linux/rpm/VirtualBox.tmpl.spec
@@ -80,7 +80,9 @@ mv *.r0 $RPM_BUILD_ROOT/usr/lib/virtualbox
 mv *.rel $RPM_BUILD_ROOT/usr/lib/virtualbox || true
 mv VBoxNetDHCP $RPM_BUILD_ROOT/usr/lib/virtualbox
 mv VBoxNetAdpCtl $RPM_BUILD_ROOT/usr/lib/virtualbox
-mv VBoxVolInfo $RPM_BUILD_ROOT/usr/lib/virtualbox
+if [ -f VBoxVolInfo ]; then
+  mv VBoxVolInfo $RPM_BUILD_ROOT/usr/lib/virtualbox
+fi
 mv VBoxXPCOMIPCD $RPM_BUILD_ROOT/usr/lib/virtualbox
 mv components $RPM_BUILD_ROOT/usr/lib/virtualbox/components
 mv *.so $RPM_BUILD_ROOT/usr/lib/virtualbox
@@ -119,6 +121,9 @@ if %WEBSVC%; then
 fi
 for i in VBoxSDL VirtualBox VBoxHeadless VBoxNetDHCP VBoxNetAdpCtl VBoxVolInfo; do
   chmod 4511 $RPM_BUILD_ROOT/usr/lib/virtualbox/$i; done
+if [ -f $RPM_BUILD_ROOT/usr/lib/virtualbox/VBoxVolInfo ]; then
+  chmod 4511 $RPM_BUILD_ROOT/usr/lib/virtualbox/VBoxVolInfo
+fi
 if [ -d ExtensionPacks/VNC ]; then
   mv ExtensionPacks/VNC $RPM_BUILD_ROOT/usr/lib/virtualbox/ExtensionPacks
 fi
